refactor(savedtrips): move info-tile handling out of SavedTripsViewModel
SavedTripsViewModel sits at detekt's LargeClass limit, so the next handler added to
it fails the build. The info tiles are the most self-contained thing in the class:
their own manager, their own analytics event, no overlap with saved trips or park
and ride. They move to SavedTripsInfoTiles.kt as extension functions, which keeps
the call sites in onEvent reading exactly as they did as members.

The members those functions need (analytics, ioDispatcher, infoTileManager,
platformOps, updateUiState) widen from private to internal. They stay inside the
module.

onParkRideCardClick also split into onParkRideCardExpanded and
onParkRideCardCollapsed, which puts it back under LongMethod. The two branches are
kept separate rather than merged with the set operation flipped: collapsing has no
"spots have arrived, stop showing loading" arm, because nothing is fetched on
collapse.

No behaviour change.
